backend/requirements.txt
fastapi
uvicorn

sqlalchemy
psycopg2-binary    # PostgreSQL driver

pandas
requests
beautifulsoup4
pyyaml
cryptography
python-dotenv
